Extension class (and login) is replaced with ApkFile
Removed too exact XML structure copying in the model. We do not need to
follow XML structure so word-for-word. In the model we can assume that
each APK file could have extensions. If there are extensions it will
correcpond to one XML tag, if there are no then it will correspond to
anothe XML tag. It is OK.